fsm-automation/requirements.txt
playwright>=1.40.0
pytest>=7.4.0
pytest-playwright>=0.4.0
allure-pytest>=2.13.0
python-dotenv>=1.0.0
pydantic>=2.5.0
python-docx>=1.1.0
